Address PR #36 review: viz overlay kwargs, test reorg, cleanup
- plot_signal_overlay: add optional reference / highlight_mask / highlight_spans kwargs and fold the manual overlay code out of plot_01 and plot_02 examples
- tests/asr: redistribute test_coverage.py and test_robustness.py into the existing per-module suites (test_estimator/adaptive/juggler) by the module each test exercises, and remove the purpose-named files
- remove docs/changes/devel/36.bugfix.rst (memory-bounding note folded into 36.feature.rst, since it is part of the new feature) and the dev-only tests/parity/matlab_reference/generate_rasr_reference.m
