+++ docs/changes/devel/36.feature.rst
@@ -9,4 +9,7 @@ sample-wise ASRDBSCAN and ASRGEV reference-selection procedures from Kim et al.
 (2025), plus the ``mne_denoise.asr.select_juggler_reference_samples`` helper,
 tests, docs, and synthetic gallery examples. Validation tooling under
 ``scripts/`` injects known burst artifacts on real EEG and reproduces headline
-figures from Chang et al. (2020), Blum et al. (2019), and Kim et al. (2025).
+figures from Chang et al. (2020), Blum et al. (2019), and Kim et al. (2025). Calibration and
+standard/adaptive processing honor an optional ``max_mem_mb`` budget through
+memory-bounded covariance aggregation, preserving the full-memory path for
+small datasets.
