+++ mne_denoise/viz/signals.py
@@ -506,6 +506,11 @@ def plot_signal_overlay(
     x_label="Time",
     y_label="Amplitude",
     title=None,
+    reference=None,
+    reference_label="Reference",
+    highlight_mask=None,
+    highlight_label="Artifact",
+    highlight_spans=None,
     show=True,
     fname=None,
 ):
@@ -536,6 +541,23 @@ def plot_signal_overlay(
         Y-axis label.
     title : str | None
         Optional custom title.
+    reference : array-like of shape (n_times,) | None
+        Optional ground-truth/reference trace overlaid on the same axis (e.g.
+        the clean signal in a simulation). Aligned to the same length and time
+        window as the before/after traces.
+    reference_label : str
+        Legend label for the reference trace.
+    highlight_mask : array-like of bool of shape (n_times,) | None
+        Optional boolean mask; samples where True are shaded with
+        ``fill_between`` to mark artifact regions.
+    highlight_label : str
+        Legend label for the ``highlight_mask`` shading.
+    highlight_spans : sequence[mapping] | None
+        Optional list of spans to shade with ``axvspan``. Each item is a mapping
+        with ``onset`` and ``duration`` (in the same units as ``times``) and
+        optional ``color``, ``alpha``, and ``label`` keys. Only the first span of
+        each distinct label receives a legend entry. Useful for annotation-based
+        repair/rejection spans.
     show : bool
         If True, display the figure.
     fname : path-like | None
@@ -610,6 +632,39 @@ def plot_signal_overlay(
         alpha=0.85,
         linewidth=1.2,
     )
+    if reference is not None:
+        ref_trace = np.asarray(reference, dtype=float).ravel()[:n_samples][mask]
+        ax.plot(
+            time_axis,
+            ref_trace,
+            color="C2",
+            label=reference_label,
+            alpha=0.8,
+            linewidth=1.0,
+        )
+    if highlight_mask is not None:
+        flagged = np.asarray(highlight_mask, dtype=bool).ravel()[:n_samples][mask]
+        ax.fill_between(
+            time_axis,
+            *ax.get_ylim(),
+            where=flagged,
+            color="C3",
+            alpha=0.12,
+            label=highlight_label,
+        )
+    if highlight_spans:
+        seen_labels = set()
+        for span in highlight_spans:
+            label = span.get("label")
+            legend_label = label if label not in seen_labels else None
+            seen_labels.add(label)
+            ax.axvspan(
+                span["onset"],
+                span["onset"] + span.get("duration", 0.0),
+                color=span.get("color", "C3"),
+                alpha=span.get("alpha", 0.12),
+                label=legend_label,
+            )
     ax.set_xlabel(x_label)
     ax.set_ylabel(y_label)
     ax.set_title(title or "Signal Overlay Comparison")
