+++ tests/asr/test_juggler.py
@@ -13,6 +13,8 @@ import pytest
 
 from mne_denoise.asr import JugglerASR, select_juggler_reference_samples
 
+SFREQ = 250.0
+
 
 def _make_synthetic_eeg(
     sfreq: float = 250.0,
@@ -296,3 +298,174 @@ def test_juggler_dbscan_deterministic():
     np.testing.assert_array_equal(
         asr1.get_calibration_mask(), asr2.get_calibration_mask()
     )
+
+
+# ===========================================================================
+# Tests relocated from former test_coverage.py / test_robustness.py (PR #36).
+# Grouped here by the module they exercise.
+# ===========================================================================
+
+
+def _eeg(n_channels=8, n_times=8000, seed=0, bursts=5):
+    rng = np.random.default_rng(seed)
+    t = np.arange(n_times) / SFREQ
+    X = np.zeros((n_channels, n_times))
+    for c in range(n_channels):
+        X[c] = 0.6 * np.sin(2 * np.pi * 10 * t + rng.uniform(0, 6.28)) + (
+            0.05 * rng.standard_normal(n_times)
+        )
+    for s in np.linspace(800, n_times - 500, bursts).astype(int):
+        spatial = rng.standard_normal(n_channels)
+        spatial /= np.linalg.norm(spatial)
+        X[:, s : s + 150] += 10.0 * np.outer(spatial, rng.standard_normal(150))
+    return X
+
+
+def _make_clean_eeg(
+    n_channels: int = 16,
+    duration_s: float = 40.0,
+    sfreq: float = 250.0,
+    seed: int = 11,
+) -> np.ndarray:
+    rng = np.random.default_rng(seed)
+    n = int(sfreq * duration_s)
+    t = np.arange(n) / sfreq
+    data = np.zeros((n_channels, n), dtype=np.float64)
+    for ch in range(n_channels):
+        phase = rng.uniform(0, 2 * np.pi)
+        data[ch] = 0.6 * np.sin(2 * np.pi * 10.0 * t + phase) + 0.15 * np.sin(
+            2 * np.pi * 6.5 * t + phase * 0.8
+        )
+    data += 0.05 * rng.standard_normal(data.shape)
+    return data
+
+
+def test_juggler_to_annotations_calibration_ok():
+    mne = pytest.importorskip("mne")
+    j = JugglerASR(
+        sfreq=SFREQ, cutoff=20.0, strategy="dbscan", picks=None, verbose=False
+    )
+    j.fit_transform(_eeg())
+    ann = j.to_annotations("calibration")
+    assert isinstance(ann, mne.Annotations)
+
+
+def test_juggler_min_reference_fraction_floor_raises():
+    from mne_denoise.asr import select_juggler_reference_samples
+
+    X = _eeg(bursts=40)  # heavily contaminated
+    with pytest.raises(RuntimeError, match="retained too little"):
+        select_juggler_reference_samples(
+            X, SFREQ, strategy="gev", min_reference_fraction=0.99
+        )
+
+
+def test_juggler_evoked_calibration_raises():
+    mne = pytest.importorskip("mne")
+    info = mne.create_info([f"EEG{i:02d}" for i in range(8)], SFREQ, "eeg")
+    evoked = mne.EvokedArray(_eeg(n_times=2000) * 1e-6, info, tmin=0.0, verbose=False)
+    with pytest.raises(ValueError, match="Evoked"):
+        JugglerASR(sfreq=SFREQ, picks=None, verbose=False).fit(evoked)
+
+
+def test_juggler_calibration_mask_shape_raises():
+    with pytest.raises(ValueError, match="calibration_mask must have shape"):
+        JugglerASR(sfreq=SFREQ, picks=None, verbose=False).fit(
+            _eeg(), calibration_mask=np.ones(10, dtype=bool)
+        )
+
+
+def test_juggler_dbscan_bad_eps_string_raises():
+    from mne_denoise.asr import select_juggler_reference_samples
+
+    with pytest.raises(ValueError, match="dbscan_eps"):
+        select_juggler_reference_samples(
+            _eeg(), SFREQ, strategy="dbscan", dbscan_eps="bogus"
+        )
+
+
+def test_juggler_dbscan_bad_min_samples_string_raises():
+    from mne_denoise.asr import select_juggler_reference_samples
+
+    with pytest.raises(ValueError, match="dbscan_min_samples"):
+        select_juggler_reference_samples(
+            _eeg(), SFREQ, strategy="dbscan", dbscan_min_samples="bogus"
+        )
+
+
+def test_juggler_dbscan_numeric_eps_fallback():
+    from mne_denoise.asr import select_juggler_reference_samples
+
+    _, mask, _ = select_juggler_reference_samples(
+        _eeg(bursts=8), SFREQ, strategy="dbscan", dbscan_eps=0.0
+    )
+    assert mask.dtype == bool
+
+
+def test_juggler_dbscan_min_samples_numeric():
+    from mne_denoise.asr import select_juggler_reference_samples
+
+    X = _eeg(bursts=8)
+    _, m1, _ = select_juggler_reference_samples(
+        X, SFREQ, strategy="dbscan", dbscan_min_samples=0.05
+    )
+    _, m2, _ = select_juggler_reference_samples(
+        X, SFREQ, strategy="dbscan", dbscan_min_samples=5
+    )
+    assert m1.dtype == bool and m2.dtype == bool
+
+
+def test_select_juggler_reference_samples_param_guards():
+    from mne_denoise.asr import select_juggler_reference_samples
+
+    X = _eeg()
+    with pytest.raises(ValueError, match="strategy must be"):
+        select_juggler_reference_samples(X, SFREQ, strategy="bogus")
+    with pytest.raises(ValueError, match="dbscan_top_k"):
+        select_juggler_reference_samples(X, SFREQ, dbscan_top_k=0)
+    with pytest.raises(ValueError, match="gev_grid_size"):
+        select_juggler_reference_samples(X, SFREQ, strategy="gev", gev_grid_size=8)
+    with pytest.raises(ValueError, match="min_reference_fraction"):
+        select_juggler_reference_samples(X, SFREQ, min_reference_fraction=1.5)
+
+
+@pytest.mark.parametrize(
+    "kwargs,msg",
+    [
+        ({"dbscan_top_k": 0}, "dbscan_top_k"),
+        ({"gev_grid_size": 8}, "gev_grid_size"),
+        ({"min_reference_fraction": 1.5}, "min_reference_fraction"),
+    ],
+)
+def test_juggler_constructor_param_guards(kwargs, msg):
+    with pytest.raises(ValueError, match=msg):
+        JugglerASR(sfreq=SFREQ, picks=None, verbose=False, **kwargs).fit(_eeg())
+
+
+@pytest.mark.parametrize("eps_multiplier", [0.5, 1.0, 2.0])
+def test_juggler_dbscan_eps_parametric(eps_multiplier):
+    """Larger eps should retain at least as many samples as smaller eps."""
+    clean = _make_clean_eeg(duration_s=60.0)
+    dirty, _ = _inject_bursts(clean, n_bursts=10)
+    # Get auto eps first
+    base = JugglerASR(
+        sfreq=250.0,
+        cutoff=20.0,
+        strategy="dbscan",
+        picks=None,
+        verbose=False,
+    )
+    base.fit(dirty)
+    auto_eps = float(base.calibration_info_["juggler_dbscan_eps"])
+    eps = auto_eps * eps_multiplier
+    asr = JugglerASR(
+        sfreq=250.0,
+        cutoff=20.0,
+        strategy="dbscan",
+        dbscan_eps=eps,
+        picks=None,
+        verbose=False,
+    )
+    asr.fit(dirty)
+    rf = float(asr.calibration_info_["reference_selected_fraction"])
+    assert 0.0 < rf <= 1.0
